Split plugin index in core & node part
